// macrocomm/src/app/api/generated/model/incidentSeverity.ts
/**
 * FastAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export const IncidentSeverity = {
    Low: 'low',
    Medium: 'medium',
    High: 'high',
    Critical: 'critical'
} as const;
export type IncidentSeverity = typeof IncidentSeverity[keyof typeof IncidentSeverity];